Add initial pass at collision
I added the collision equation. I'm not entirely sure what all the
coefficients have to do with anything, and in general at a fine-grained
level, I haven't intuited with the equation is actually DOING. But... I
think I understand what colliding means: basically, after the streaming
(which is simply moving particles around to other lattice sites), each
lattice site tries to reach equilibrium. This happens in the physical
world when molecules in a fluid crash together (collide!) and exchange
momentum and energy. Equilibrium is highly dependent on the viscosity
of the fluid you're simulating. According to the paper, a substance like
honey will reach equilibrium slower than water. This seems counter
intuitive: if I jiggle a bowl of water, it keeps moving long after I
stop with the jigglage. If I jiggle a bowl of honey, it stops moving
very quickly. I dunno. But... what do they mean by equilibrium? Thermal?
Kinetic? ???

Good first pass, though. Learned about the dot product, got a high-level
intuition of what collision means in the LBM. Next up, I think I need to
introduce barriers and implement bounce-back. Right now, I can't tell if
anything is flowing, and all the densities are uniform because there's
nothing in the way. I also haven't constrained the top/bottom to act as
walls in a wind tunnel. As far as I can tell, I've implemented a void
that may or may not have still air in it. Once I've got barriers in
place, I can implement a "click-to-readout" feature, where I can click a
lattice site and see a readout of the density at that lattice site.
After that, color!