Stable?
I got it to stabilize to "equilibrium" by changing the viscosity value
in the omega equation to be higher that I thought it should be. The
example I'm following shows it going from 0.005 to 0.200. I changed mine
to be about 2+ and things seem like their stable. however, the density
for each site hovers around 4... shouldn't it be 1?