It's the collide...
If I comment out the collide step, I see the velocity equal the expected
values. I'm pretty sure my densities are not equilibrating correctly,
still. No clue why not... The reason I think this is because I was
having trouble with them before and I fudged the value which made it act
a little nicer. Also the macro velocities are apparently derived from
the displacement densities of the site. So if those aren't correct, the
velocity won't be correct..