Sheesh
So one majore problem I had was that I was streaming SW densities to W
densities, instead of W -> W. :facepalm: That fixes the problem I was
having with the density not settling to 1 at rest. It also fixes the
problem I was having where the X velocity was ALWAYS 1.5 if I gave it a
rightward velocity to start. A new problem has shown up - if I do init
the fluid to have a rightward flow of 1, the density settles to NaN. I
don't know why this happens, but it may be worth looking
at/tweaking/comparing the streaming step since it has such a big impact
on the collision step. I actually think we can say the collision step is
good to go, since it works for a fluid at rest. Also consider involving
the time scale. It might be interesting to slow down the time step so
that it's easier to visualize the lattice sites coming into equilibrium.