Not a Not a Number
A lot has happened. I revamped streaming to account for wrapping edges.
Then I set the flows around the edges manually and undid all that work.
It's simpler than it was, but I still feel like something isn't right. I
have the overall flow speed set to 0.120, but the actual speed at the
sites is set to something much lower. Density looks... ok, but is still
higher than the 1 I think it should be. If I set the viscosity in the
site (OMEGA const) to something less than about 0.2, the thing freaks
out. Used to give me a NaN, but that might be fixed with adding edges.

I would love to set up some sort of visual debugging. It would help me
intuitively and also potentially help solve the problem. I'm thinking:
1. boxes around each of the 9 discretized densities of each site.
2. color those boxes. Not sure how or with what info - making sure they
   add up to and overall density of 1 might be good
3. Arrows pointing from each of those boxes to the site the particles
   will stream to
4. incorporate the javascript/dev tools debugger to step through OR
5. create some sort of time controls to slow/speed things down/up.