D1Q3
Simplified and reduced to a single dimension. The velocity and densities
seem to be spot on. One thing I changed was setting the end site
(rightmost) to flow w/ the same speed as the beginning site. I did this
because I noticed the density "piling up" in the last site to something
crazy like 7.something. I figured the fluid was getting "trapped" and
the pressure was increasing. It's cool to be able to think about the
simulation as a fluid. I'll go back to the 2d version and check this,
but I think I was doing the right thing with the end sites.