Math and stuff.
I expanded the collision equation on paper to compare with the code I'm
referencing and it all looked good. Then I looked at the streaming code
and double checked I wasn't overwiting any values. I noticed that it
only has a problem when streaming to N/S sites. I noticed this because
the simulation is fine when there's a single row available to the fluid
(excluding the `setEquil` rows (important!)). So after confirming that
by only streaming to the E/W sites with several rows available to the
sim, I thought that the boundaries may be acting too much like walls.
They're just set to streams of fluid with a constant flow (or no flow).
So I took a look at where I was looping over the collide function for
each site and noticed I don't collide AT ALL with the `setEquil`
rows/columns. For streaming and reflection, I don't know if it's easy to
include these rows/cols, but with collision, there's no harm in it, as
collision occurs w/in the site. So I changed the loop to include 0 and
gW - 1 & gH - 1, and we have a nice sim with no boundaries at any
reasonable flow (0 - 0.120).

EDIT: not ANY reasonable flow. It seems to have trouble with like..
0.0020 for some reason?!?