Solved
So after reading the wikipedia article, I kept seeing the south-pointing
vector as (0, -1). I had it as (0, 1), as in a grid framework. I think
that south is still (0, 1) in the grid, but having the allowed, south
pointing vector of a site be (0, -1) seems to matter. I swapped all the
north and south (including angled directions: ne se nw sw) directions
and got great results even with a higher flow speed and low viscosity! I
don't know why it has to be (0, -1), but I assume the math of the
collision step works out that way.

TODO: figure out why south vector of a site has to be (0, -1)